+++ client/pages/blog/blogwrite.tsx
@@ -1,13 +1,13 @@
 import tw from "tailwind-styled-components";
-import Header from "../src/components/Page/UserHome/Header";
-import Navbar from "../src/components/Nav/NavBar";
-import UserProfileImage from "../src/components/Page/UserHome/UserProfileImage";
-import UserProfile from "../src/components/Page/UserHome/UserProfile";
-import Followig from "../src/components/Page/UserHome/Following";
-import TodayCount from "../src/components/ViewLikeWrite/TodayCount";
-import { Content, BlogContent, UserContent, UserInfo } from "./blog";
-import Footer from "../src/components/Footer/Footer";
-import BlogWriteComponent from "../src/components/Page/BlogWrite/BWComponent";
+import Header from "../../src/components/Page/UserHome/Header";
+import Navbar from "../../src/components/Nav/NavBar";
+import UserProfileImage from "../../src/components/Page/UserHome/UserProfileImage";
+import UserProfile from "../../src/components/Page/UserHome/UserProfile";
+import Followig from "../../src/components/Page/UserHome/Following";
+import TodayCount from "../../src/components/ViewLikeWrite/TodayCount";
+import { Content, BlogContent, UserContent, UserInfo } from "./[userId]";
+import Footer from "../../src/components/Footer/Footer";
+import BlogWriteComponent from "../../src/components/Page/BlogWrite/BWComponent";
 
 export const EditorContent = tw.div`w-[43rem] m-auto`;
